+++ lib/page/setting/screen/change_log_page.dart
@@ -1,9 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_markdown/flutter_markdown.dart';
-
-import '../widgets/bookify_ads.dart';
-// import 'package:startapp_sdk/startapp.dart';
+import '../widgets/snapyt_ads.dart';
 
 class ChangeLogPage extends StatefulWidget {
   const ChangeLogPage({super.key});
@@ -15,22 +13,9 @@ class ChangeLogPage extends StatefulWidget {
 class ChangeLogPageState extends State<ChangeLogPage> {
   String markdownData = '';
 
-  // var startApp = StartAppSdk();
-  // StartAppBannerAd? bannerAds;
-
-  // loadBannerAds() {
-  //   //! startApp.setTestAdsEnabled(true);
-  //   startApp.loadBannerAd(StartAppBannerType.BANNER).then((value) {
-  //     setState(() {
-  //       bannerAds = value;
-  //     });
-  //   });
-  // }
-
   @override
   void initState() {
     super.initState();
-    // loadBannerAds();
     loadMarkdown();
   }
 
@@ -47,9 +32,6 @@ class ChangeLogPageState extends State<ChangeLogPage> {
       appBar: AppBar(
         title: const Text('Chanage History'),
       ),
-      // bottomNavigationBar: bannerAds != null
-      //     ? SizedBox(height: 60, child: StartAppBanner(bannerAds!))
-      //     : const SizedBox(),
       bottomNavigationBar: const BookifyAds(
         apiUrl: 'https://gokeihub.github.io/bookify_api/ads/changelog.json',
       ),